requirements.txt
bcrypt==3.2.2
cffi==1.15.1
cryptography==37.0.4
paramiko==2.11.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.20.0
six==1.16.0
